chore(tasks): Add logging for when tasks take a long time to run (#3919)
* chore(tasks): Add logging for when tasks take a long time to run

This change adds logging for when a tasks `execute` method takes a long time to run (over 30s).
Since tasks have a pretty sensitive ack timeout, having a long running taks might be indicative of a problem to come.

We have metrics for task invocation times as well, but once an operator knows that some task takes a long time, it's useful to find it in logs to see what it was doing for further investigation

* fixup! chore(tasks): Add logging for when tasks take a long time to run

* fixup! fixup! chore(tasks): Add logging for when tasks take a long time to run