feat(stages): Add stage to monitor specific execution(s) (#3902)
* feat(stages): Add stage to monitor specific execution(s)

This change adds a `MonitorPipelineStage` stage that allows one to monitor a given set of executions, similar to how `PipelineStage` starts and monitors that pipelins.
The difference is that this stage doesn't start any pipelines, it simply performs the monitoring.

Ideally, I would use the `MonitorPipelineStage` as a synthetic child of `PipelineStage` but that requires a bit more finesse and refactor that I don't think is worth it right now.

I will also be making corresponding `deck` changes to visualize the stage execution

* fixup! feat(stages): Add stage to monitor specific execution(s)

* fixup! fixup! feat(stages): Add stage to monitor specific execution(s)

* error messages

* fixup! error messages

* fixup! fixup! error messages

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>