feat(disable): add an optional waitForDisabledServerGroup task (#3943)
* feat(disable): add an optional waitForDisabledServerGroup task

When we disable a server group, we disable and also take instances down, but only wait for instances to become down. This means that it is possible to finish a DisableServerGroupStage successfully with down instances but a server group that does not yet appear disabled. This creates poor interactions, for instance if the stage that follows is a resize.